Pythonproject/requirements.txt
altair==5.3.0
amqp==5.3.1
attrs==23.2.0
billiard==4.2.1
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==3.0.2
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-Cors==4.0.0
Flask-Session==0.8.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
msgspec==0.18.6
mysql-connector-python==8.3.0
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prompt_toolkit==3.0.48
protobuf==4.25.3
pyarrow==16.0.0
PyAutoGUI==0.9.54
pydeck==0.9.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
redis==5.2.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
streamlit==1.34.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
waitress==3.0.2
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
zstandard==0.23.0
